+++ src/scripts/libs/font-color-contrast-1.1.1/cssNamedColors.d.ts
@@ -1,0 +1,5 @@
+export declare const cssNamedColors: {
+    name: string;
+    hex: string;
+}[];
+//# sourceMappingURL=cssNamedColors.d.ts.map